+++ Example/LocalizationDemo/SecondViewController.m
@@ -11,6 +11,7 @@
 @interface SecondViewController ()
 @property (nonatomic, weak) IBOutlet UILabel *countLabel;
 @property (nonatomic, weak) IBOutlet UIStepper *stepper;
+@property (nonatomic, weak) IBOutlet UISegmentedControl *bundleSelector;
 @property (nonatomic, strong) NSNumberFormatter *numberFormatter;
 @end
 
@@ -24,12 +25,34 @@
 
     DPFormattedValue *numberValue = [DPFormattedValue formattedValueWithValue:@(self.stepper.value) formatter:self.numberFormatter];
     [self.countLabel updateAutolocalizationArguments:@[numberValue]];
-}
 
+    self.bundleSelector.selectedSegmentIndex = ([[DPLocalizationManager currentManager].defaultBundle isEqual:[NSBundle mainBundle]]) ? 0 : 1;
+}
 
 - (IBAction)stepperValueChanged:(id)sender {
     DPFormattedValue *numberValue = [DPFormattedValue formattedValueWithValue:@(self.stepper.value) formatter:self.numberFormatter];
     [self.countLabel updateAutolocalizationArguments:@[numberValue]];
+}
+
+- (IBAction)bundleDidChange:(id)sender {
+    switch (self.bundleSelector.selectedSegmentIndex) {
+        case 0:
+            [DPLocalizationManager currentManager].defaultBundle = nil;
+            break;
+
+        case 1: {
+            NSString *bundlePath = [[[NSBundle mainBundle] resourcePath] stringByAppendingPathComponent:@"CustomBundle"];
+            [DPLocalizationManager currentManager].defaultBundle = [[NSBundle alloc] initWithPath:bundlePath];
+            break;
+        }
+
+        default:
+            break;
+    }
+
+    NSLog(@"Preffered language: %@", [[DPLocalizationManager currentManager].defaultBundle preferredLanguage]);
+    NSLog(@"Selected language: %@", dp_get_current_language());
+    NSLog(@"Supported language: %@", [[DPLocalizationManager currentManager].defaultBundle supportedLanguages]);
 }
 
 @end