+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # DPLocalization CHANGELOG
 
+## 1.8
+- Fixed Set current language to 'nil' before reading
+- Added defaultBundle property. ".strings" resources can be stored outside mainBundle.
+
 ## 1.7
 - Added Plural rules (NSNumber - integer only, DPFormattedNumberValue - decimal)
 - Added -[DPLocalizationManager localizedStringForKey: table: arguments:]